Initial support of generated authorized_keys
This commit provides the initial support and Travis tests for
dynamically generating an authorized_keys file given a user/role and a
list of GitHub usernames. The SSH keys for the GitHub users are applied
to the user/role account.

The current testing in Travis does the following:
 - Use geerlinguy's testing harness
 - Create a user (jhriv), grab their SSH key(s) from GitHub, and ensure
the first key is present in the generated authorized_keys file created
by the role